requirements.txt
pygame~=2.6.1
colorama~=0.4.6
termcolor~=2.5.0
pytest
pillow~=11.0.0